Renamed the docker build flags object
